+++ lib/src/commands/fetch_tags_command.dart
@@ -6,9 +6,10 @@ class FetchTagsCommand extends GitCommand<List<String>> {
   FetchTagsCommand(super.runner);
 
   @override
-  Future<List<String>> execute(String directory) async {
-    final result =
-        await runner.run('git', ['tag', '-l'], workingDirectory: directory);
+  Future<List<String>> execute(String directory,
+      {bool streamOutput = false}) async {
+    final result = await runner.run('git', ['tag', '-l'],
+        workingDirectory: directory, streamOutput: streamOutput);
     evaluateProcessResult(result);
     return RwGitParser.parseGitStdoutBasedOnNewLine(result.stdout.toString());
   }