+++ lib/src/commands/is_git_repository_command.dart
@@ -4,10 +4,10 @@ class IsGitRepositoryCommand extends GitCommand<bool> {
   IsGitRepositoryCommand(super.runner);
 
   @override
-  Future<bool> execute(String directory) async {
+  Future<bool> execute(String directory, {bool streamOutput = false}) async {
     final result = await runner.run(
         'git', ['rev-parse', '--is-inside-work-tree'],
-        workingDirectory: directory);
+        workingDirectory: directory, streamOutput: streamOutput);
     // don't evaluateProcessResult because if it's not a git repo, it will exit non-zero and we just return false
     return result.exitCode == 0 &&
         result.stdout.toString().toLowerCase().trim() == "true";