+++ lib/src/commands/mega_commit_command.dart
@@ -8,11 +8,12 @@ class MegaCommitCheckCommand extends GitCommand<List<String>> {
   MegaCommitCheckCommand(super.runner, {this.thresholdLinesChanged = 500});
 
   @override
-  Future<List<String>> execute(String directory) async {
+  Future<List<String>> execute(String directory,
+      {bool streamOutput = false}) async {
     // git log --shortstat returns the stat per commit which we can parse to find mega-commits
     final result = await runner.run(
         'git', ['log', '--shortstat', '--format=oneline'],
-        workingDirectory: directory);
+        workingDirectory: directory, streamOutput: streamOutput);
     evaluateProcessResult(result);
 
     // the parser logic will be handled outside, but for now we just return the raw lines to be processed